[tune] Move experiment checkpoint/state management into separate file
